requirements.txt
# Telegram Bot Framework
python-telegram-bot[all]>=20.7

# Database
sqlalchemy[asyncio]>=2.0
alembic>=1.13.0
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Redis
aioredis>=2.0.1
redis>=5.0.1

# Task Scheduling
apscheduler>=3.10.4

# HTTP Client
aiohttp>=3.9.1

# Image Processing
Pillow>=10.2.0

# Audio Processing
pydub>=0.25.1

# Language Detection
langdetect>=1.0.9

# OpenAI Integration
openai>=1.10.0

# QR Code Generation
qrcode[pil]>=7.4.2

# Environment Variables
python-dotenv>=1.0.0

# Timezone Support
pytz>=2024.1

# Internationalization
Babel>=2.14.0

# System Monitoring
psutil>=5.9.8

# Cryptography
cryptography>=42.0.0

# Data Validation
pydantic>=2.6.0

# Async Utilities
anyio>=4.2.0
asyncio-throttle>=1.0.2

# Text Processing
ftfy>=6.1.3
emoji>=2.10.1

# HTML/Markdown Parsing
beautifulsoup4>=4.12.3
markdown>=3.5.2

# Templating
jinja2>=3.1.3

# Logging
loguru>=0.7.2

# Date/Time
python-dateutil>=2.9.0
arrow>=1.3.0

# Hashing
bcrypt>=4.1.2

# JSON Handling
orjson>=3.9.14
ujson>=5.9.0

# Rate Limiting
limits>=3.8.0

# FFMPEG Wrapper
ffmpeg-python>=0.2.0

# Charts & Graphs
matplotlib>=3.8.2

# Random Word Generation
randomwords>=0.4.0

# Trivia / Games
aiofiles>=23.2.1

# Spell Check
pyspellchecker>=0.8.1

# Country/Flag Data
pycountry>=23.12.11

# URL Shortener Client
pyshorteners>=1.0.1

# Misc Utilities
humanize>=4.9.0
cachetools>=5.3.2
tenacity>=8.2.3
